+++ src/tests/parse/test_unions.h
@@ -59,7 +59,7 @@ TEST test_write_union_to_json(void) {
   content = (char *)malloc(sz + 1);
   if (!content)
     FAILm("OOM");
-  fread(content, 1, sz, tmp);
+  fread(content, 1, (size_t)sz, tmp);
   content[sz] = 0;
 
   /* Check for switch on tag */
@@ -104,7 +104,7 @@ TEST test_write_union_from_json_object(void) {
   content = (char *)malloc(sz + 1);
   if (!content)
     FAILm("OOM");
-  fread(content, 1, sz, tmp);
+  fread(content, 1, (size_t)sz, tmp);
   content[sz] = 0;
 
   /* Check malloc */
@@ -147,7 +147,7 @@ TEST test_write_union_from_json(void) {
   content = (char *)malloc(sz + 1);
   if (!content)
     FAILm("OOM");
-  fread(content, 1, sz, tmp);
+  fread(content, 1, (size_t)sz, tmp);
   content[sz] = 0;
 
   ASSERT(strstr(content, "json_parse_string"));
@@ -188,7 +188,7 @@ TEST test_write_union_cleanup(void) {
   content = (char *)malloc(sz + 1);
   if (!content)
     FAILm("OOM");
-  fread(content, 1, sz, tmp);
+  fread(content, 1, (size_t)sz, tmp);
   content[sz] = 0;
 
   /* Check switch */
